requirement.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.3
django-admin-interface==0.28.6
django-colorfield==0.11.0
django-import-export==3.3.7
django-shopping-cart==0.1
django-tinymce==3.7.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.2.0
python-slugify==8.0.4
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
typing_extensions==4.10.0
tzdata==2024.1
xlrd==2.0.1
xlwt==1.3.0
